lake/requirements.txt
appdirs==1.4.3
arrow==0.15.5
astroid==2.3.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
networkx==2.4
numpy==1.18.1
pandas==1.0.1
pathspec==0.7.0
Pillow==7.0.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
